Moved TypeDescriptor to type package
